rename project name test to CinephileLog
