+++ src/main/java/com/CinephileLog/CinephileLogApplication.java
@@ -1,15 +1,15 @@
-package com.test;
+package com.CinephileLog;
 
 import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-@MapperScan("com.test.mapper")
-public class TestApplication {
+@MapperScan("com.CinephileLog.mapper")
+public class CinephileLogApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(TestApplication.class, args);
+        SpringApplication.run(CinephileLogApplication.class, args);
     }
 
 }
