+++ src/main/java/com/CinephileLog/domain/Test.java
@@ -1,4 +1,4 @@
-package com.test.domain;
+package com.CinephileLog.domain;
 
 import lombok.Getter;
 import lombok.Setter;
