+++ src/main/java/com/CinephileLog/dto/TestDTO.java
@@ -1,4 +1,4 @@
-package com.test.dto;
+package com.CinephileLog.dto;
 
 import lombok.Data;
