+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.application.name=test
+spring.application.name=CinephileLog
 
 server.port=8080
 spring.datasource.driver-class-name=net.sf.log4jdbc.sql.jdbcapi.DriverSpy
