+++ src/main/resources/mappers/test.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.test.mapper.TestMapper">
+<mapper namespace="com.CinephileLog.mapper.TestMapper">
     <select id="test" resultType="int">
         SELECT 3
     </select>
