+++ src/test/java/com/test/TestApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class TestApplicationTests {
+class CinephileLogApplicationTests {
 
     @Test
     void contextLoads() {
